refactor yam_loader.py into validator.py
